requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
colorclass==2.2.0
Django==3.2.3
django-filter==2.4.0
djangorestframework==3.12.4
docopt==0.6.2
idna==3.1
packaging==20.9
Pillow==8.2.0
pip-upgrader==1.4.15
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
terminaltables==3.1.0
urllib3==1.26.4
